revert to int default on lazy segtrees

+++ lazy_propagation_add_findmin.cpp
@@ -43,22 +43,22 @@ template<typename T> using ordered_map = tree<T, int, less<T>, rb_tree_tag, tree
 //Lazy Propagation on Increment Modifications, findmin queries
 
 #define N 10000
-ll t[N << 2];
-ll lazy[N << 2];
+int t[N << 2];
+int lazy[N << 2];
 
-void build(ll v, ll tl, ll tr, int arr[]){
+void build(int v, int tl, int tr, int arr[]){
     if(tl == tr){
         t[v] = arr[tl];
         return;
     }
     lazy[v] = 0;
-    ll tm = (tl+tr)>>1;
+    int tm = (tl+tr)>>1;
     build(v<<1,tl,tm,arr);
     build(v<<1|1,tm+1,tr,arr);
     t[v] = min(t[v<<1] , t[v<<1|1]);
 }
 
-void push(ll v){
+void push(int v){
     if(lazy[v]){
         lazy[v<<1] += lazy[v];
         lazy[v<<1|1] += lazy[v];
@@ -68,7 +68,7 @@ void push(ll v){
     }
 }
 
-void modify(ll v, ll tl, ll tr, ll l, ll r, ll inc){
+void modify(int v, int tl, int tr, int l, int r, int inc){
     if(l > r) return;
     if(tl == l && tr == r){
         t[v] += inc;
@@ -76,7 +76,7 @@ void modify(ll v, ll tl, ll tr, ll l, ll r, ll inc){
         return;
     }
     push(v);
-    ll tm = (tl+tr)>>1;
+    int tm = (tl+tr)>>1;
     modify(v<<1,tl,tm,l,min(tm,r),inc);
     modify(v<<1|1,tm+1,tr,max(l,tm+1),r,inc);
     t[v] = min(t[v<<1],t[v<<1|1]);
